+++ CHANGELOG.md
@@ -15,6 +15,11 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Fixes a styling issue on the range `DatePicker` for the month and year types
   where the in-range selected date values did not have the correct DS colors.
+- Fixes the desktop font size for the `"heading5"` variant in the `Heading` component.
+
+### Updates
+
+- Updates focus ring color in `Notification` and `Breadcrumbs` to match color of text.
 
 ## 2.1.5 (February 5, 2024)